Create repositories for Redis cache and deserialization of models
Plus a Thor task that can cache the JIRA data into Redis
