+++ app/models/project.rb
@@ -8,4 +8,8 @@ class Project < ActiveModelSerializers::Model
   def self.from_jira(jira_project)
     new(jira_project.key, jira_project.name, jira_project.attrs['avatarUrls'])
   end
+
+  def self.from_cache(json)
+    new(json['key'], json['name'], json['avatars'])
+  end
 end
