+++ app/models/team.rb
@@ -8,6 +8,12 @@ class Team < ActiveModelSerializers::Model
     @name, @board_id, @subteam = name, board_id, subteam
   end
 
+  def self.from_cache(json)
+    team = new(json['name'], json['board_id'], json['subteam'])
+    team.project = Project.from_cache(json['project'])
+    team
+  end
+
   def_delegator :@project, :avatars
   def_delegator :@project, :name, :project_name
   def_delegator :@project, :key, :project_key
