+++ app/repositories/cache/redis_client.rb
@@ -1,0 +1,13 @@
+require 'redis'
+
+module Cache
+  class RedisClient
+    extend Forwardable
+
+    def initialize
+      @client = Redis.new(host: 'redis')
+    end
+
+    def_delegators :@client, :get, :set, :keys
+  end
+end
