--- app/repositories/epic_repository.rb
@@ -1,10 +1,0 @@
-class EpicRepository
-  def initialize(jira_client)
-    @records = {}
-    @client = jira_client
-  end
-
-  def find(key)
-    @records[key] ||= Epic.from_jira(@client.Issue.find(key))
-  end
-end
