+++ app/repositories/jira/epic_repository.rb
@@ -1,0 +1,7 @@
+module Jira
+  class EpicRepository < Jira::JiraRepository
+    def find(key)
+      @records[key] ||= Epic.from_jira(@client.Issue.find(key))
+    end
+  end
+end
