+++ app/repositories/jira/jira_repository.rb
@@ -1,0 +1,8 @@
+module Jira
+  class JiraRepository
+    def initialize(jira_client)
+      @records = {}
+      @client = jira_client
+    end
+  end
+end
