--- app/repositories/project_repository.rb
@@ -1,10 +1,0 @@
-class ProjectRepository
-  def initialize(jira_client)
-    @records = {}
-    @client = jira_client
-  end
-
-  def find(key)
-    @records[key] ||= Project.from_jira(@client.Project.find(key))
-  end
-end
