+++ config.yml
@@ -1,0 +1,1 @@
+:use_cached_data: true
